Java Refractor (#266)
This contains the next release of Remi, a full infrastructure refractor of the project. It is now purely made in Java to let Java users use the power of Remi and not have `suspend` functions.

The project also contains Kotlin extensions and kotlinx.serialization support for configurations for the supported storage services.

This release also implements Azure and Google Cloud Storage support.